refactor: healthController controller 패키지 안에 배치
